new stuff

+++ A3/ext2_cp.c
@@ -49,7 +49,7 @@ int main(int argc, char **argv) {
         exit(1);
     }
     
-    size_t filesize = lseek(source, 0, SEEK_END);
+    float filesize = lseek(source, 0, SEEK_END);
     int blockneeded = ceil(filesize/1024);
 	printf("%d",blockneeded);
     if (blockneeded > 12) {
